+++ lib/Steps/ChooseIntegration.ts
@@ -18,6 +18,7 @@ import { Apple } from './Integrations/Apple';
 import { SvelteKitShim } from './Integrations/SvelteKitShim';
 import { hasPackageInstalled } from '../../src/utils/package-json';
 import { Remix } from './Integrations/Remix';
+import { Android } from './Integrations/Android';
 
 let projectPackage: any = {};
 
@@ -38,6 +39,9 @@ export class ChooseIntegration extends BaseStep {
 
     let integration = null;
     switch (integrationPrompt.integration) {
+      case Integration.android:
+        integration = new Android(this._argv);
+        break;
       case Integration.cordova:
         integration = new Cordova(sanitizeUrl(this._argv));
         break;
@@ -97,7 +101,7 @@ export class ChooseIntegration extends BaseStep {
       return { integration: this._argv.integration };
     } else {
       if (this._argv.quiet) {
-        throw new Error('You need to choose a integration');
+        throw new Error('You need to choose a platform');
       }
 
       const detectedDefaultSelection = this.tryDetectingIntegration();
@@ -106,7 +110,7 @@ export class ChooseIntegration extends BaseStep {
         {
           choices: getIntegrationChoices(),
           default: detectedDefaultSelection,
-          message: 'What integration do you want to set up?',
+          message: 'What platform do you want to set up?',
           name: 'integration',
           type: 'list',
         },